Comparative study of the classification models
Abstract — This research paper has evaluated various classification models for prediction of bank telemarketing campaign results regarding the probability of the subscription of the customer to the deposit. The effectiveness of these algorithms has been evaluated by Receiving Operator Characteristic (ROC) and Cumulative Accuracy Profile (CAP) curve analysis, the accuracy of the algorithm and variance of the predictions. According to the results of the research, the best model for bank telemarketing effectiveness prediction are Random Forest and Deep Artificial Neural Network. The Logistic Regression and Naive Bayes are not as suitable as the other classification methods for this kind of problems due to the poor accuracy and overfitting issues.